+++ src/main/java/nl/esciencecenter/praline/newalign/AlignResult.java
@@ -5,10 +5,10 @@ import nl.esciencecenter.praline.aligners.AlignStep;
 import java.util.List;
 
 public class AlignResult {
-    public final float score;
+    public final double score;
     public final List<AlignStep> align;
 
-    public AlignResult(float score, List<AlignStep> align) {
+    public AlignResult(double score, List<AlignStep> align) {
         this.score = score;
         this.align = align;
     }
