+++ src/main/java/nl/esciencecenter/praline/newalign/IGapCost.java
@@ -2,5 +2,5 @@ package nl.esciencecenter.praline.newalign;
 
 public interface IGapCost {
 
-    float getGapCost(int gapSize);
+    double getGapCost(int gapSize);
 }
