+++ src/main/java/nl/esciencecenter/praline/newalign/IPositionCost.java
@@ -1,5 +1,5 @@
 package nl.esciencecenter.praline.newalign;
 
 public interface IPositionCost {
-    float cost(int posA, int posB);
+    double cost(int posA, int posB);
 }
